backend/Eduschola/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.33.0
colorama==0.4.6
comm==0.1.2
cryptography==40.0.2
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==4.0.0
django-webpack-loader==1.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.16.7
exceptiongroup==1.1.1
executing==1.2.0
faiss-cpu==1.7.4
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.23.1
ipython==8.13.0
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
langchain==0.0.220
langchainplus-sdk==0.0.19
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.25.0
oauthlib==3.2.2
openai==0.27.7
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pgvector==0.1.8
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.10
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==3.0.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.4
python3-openid==3.2.0
pytweening==1.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306;platform_system == "Windows"
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
selenium==4.10.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.17
sqlparse==0.4.4
stack-data==0.6.2
streamlit==1.24.0
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.7.0
tzdata==2022.7
tzlocal==4.3.1
urllib3==1.26.14
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
wget==3.2
whitenoise==6.4.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
